+++ src/Auto3D/SPE.py
@@ -8,10 +8,10 @@ from rdkit import Chem
 
 from Auto3D.batch_opt.model_wrapper import EnForce_ANI
 from Auto3D.batch_opt.padding import pad_from_mols
-from Auto3D.constants import EV_TO_HARTREE
 from Auto3D.model_factory import create_model, get_device
 from Auto3D.models.preflight import resolve_engine_name
 from Auto3D.torch_config import TorchConfig, configure_torch
+from Auto3D.utils.energy import set_e_hartree_from_ev
 from Auto3D.utils.logging_config import get_logger
 from Auto3D.utils.output_guard import check_output_not_input, check_output_overwrite
 from Auto3D.utils.validation import (
@@ -182,7 +182,7 @@ def calc_spe(
 
     with Chem.SDWriter(str(outpath)) as f:
         for i, mol in enumerate(mols):
-            mol.SetProp('E_hartree', str(es[i] * EV_TO_HARTREE))
+            set_e_hartree_from_ev(mol, float(es[i]))
             f.write(mol)
     return str(outpath)